requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.30
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
docker>=7.0.0
aiofiles>=23.2.0
markdown2>=2.4.10
bcrypt==3.2.2
email-validator>=2.1.0
pydantic-settings>=2.0.0 